Data_Cleaning_Project_Scripts
Built a data cleaning project. Given a dataset of lay-offs for companies throughout the world, the sql script was designed to command specific jobs. The goal of the project was to prepare the data set for data exploratory projects. To do so, the sql scripts ran were designed to remove duplicates, remove columns, standardize, data, and address Null values, or blank values.